pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "liteclaw"
version = "0.1.0"
description = "Wrapper Python para LiteRT-LM com suporte a tools"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.28.0",
    "websockets>=12.0",
    "aiohttp>=3.9",
    "python-telegram-bot>=21.0",
    "discord.py>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["liteclaw*"]

[project.scripts]
liteclaw = "liteclaw.cli:main"
